+++ R/hullplot.R
@@ -16,10 +16,13 @@
 # You should have received a copy of the GNU General Public License along
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- 
+
 hullplot <- function(x, cl, col = NULL,
   cex = 0.5, hull_lwd = 1, hull_lty = 1, solid = TRUE, alpha = .2,
   main = "Convex Cluster Hulls", ...) {
+
+  ### handle d>2 by using PCA
+  if(ncol(x)>2) x <- prcomp(x)$x
 
   ### extract clustering (keep hierarchical xICSXi structure)
   if(is(cl, "xics") || "clusters_xi" %in% names(cl)) {
@@ -33,7 +36,7 @@ hullplot <- function(x, cl, col = NULL,
   #if(is.null(col)) col <- c("#000000FF", rainbow(n=max(cl)))
   if(is.null(col)) col <- palette()
   if(max(cl)+1L > length(col)) warning("Not enough colors. Some colors will be reused.")
-  
+
   plot(x[,1:2], col = col[cl %% length(col) +1L], cex = cex, main = main, ...)
   col_poly <- adjustcolor(col, alpha.f = alpha)
   border <- col
@@ -53,12 +56,12 @@ hullplot <- function(x, cl, col = NULL,
   for(i in 1:length(ci_order)) {
 
     ### use all the points for xICSXi's hierarchical structure
-    if(is.null(clusters_xi)) { d <- x[cl==i,] 
+    if(is.null(clusters_xi)) { d <- x[cl==i,]
     } else { d <- x[cl_order[clusters_xi$start[i] : clusters_xi$end[i]],] }
 
     ch <- chull(d)
     ch <- c(ch, ch[1])
-    if(!solid) { lines(d[ch,], col = col[ci_order[i] %% length(col) +1L], lwd=hull_lwd, lty=hull_lty) 
+    if(!solid) { lines(d[ch,], col = col[ci_order[i] %% length(col) +1L], lwd=hull_lwd, lty=hull_lty)
     } else { polygon(d[ch,], col = col_poly[ci_order[i] %% length(col_poly) +1L],
                      lwd=hull_lwd, lty=hull_lty, border = border[ci_order[i] %% length(col_poly) +1L]) }
   }