+++ R/kNN.R
@@ -54,7 +54,7 @@ kNN <- function(x, k, sort = TRUE, search = "kdtree", bucketSize = 10,
     d <- t(sapply(1:nrow(id), FUN = function(i) x[i, id[i,]]))
     dimnames(d) <- list(rownames(x), 1:k)
 
-    return(list(dist = d, id = id, k = k))
+    return(structure(list(dist = d, id = id, k = k), class = "kNN"))
   }
 
   ## make sure x is numeric